:recycle: Inline planner search params handling
